requirements.txt
pillow
json
numpy
